requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
livekit-api
flask[async]
flask
flask-cors
uvicorn
firebase-admin
livekit
phidata
openai
lancedb
tantivy
pypdf
sqlalchemy
